requirements.txt
python-dotenv
notion-client
openai